requirements.txt
asgiref==3.9.1
Django==5.2.4
djangorestframework==3.16.0
setuptools==78.1.1
sqlparse==0.5.3
tzdata==2025.2
wheel==0.45.1
